backend/pyproject.toml
[project]
name = "polymock"
version = "0.1a"
description = "polymock is an copy of polymarket (pet-project)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.17.2",
    "apscheduler>=3.10.0",
    "asyncpg>=0.31.0",
    "fastapi>=0.125.0",
    "httpx>=0.28.0",
    "instructor>=1.7.0",
    "openai>=1.59.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=4.0.0,<4.1.0",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.8.0",
    "python-multipart>=0.0.21",
    "sqlalchemy>=2.0.45",
    "thefuzz>=0.22.0",
    "uvicorn[standard]>=0.38.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]
